requirements.txt
aiomysql
aiosqlite
dateparser>=1.1.0
decamelize==0.1.2
fastapi[all]>=0.63.0,<0.70
fastapi-pagination~=0.9.0
grpcio>=1.32.0,<1.42
grpcio-tools>=1.32.0,<1.42
grpc-interceptor==0.13.0
kombu>=5.2.4,<6.0
MarkupSafe==2.0.1
PyMySQL<=0.9.3,>=0.9
passlib[bcrypt]==1.7.2
pillow>=7.2.0,<8.4
protobuf>=3.15,<3.18
pydantic-sqlalchemy>=0.0.7,<1
python-jose[cryptography]==3.2.0
pytz==2021.1
redis>=3.5.3,<4.3
regex==2022.1.18
sqla-wrapper>=5.0.0,<5.6
typer>=0.4
uvicorn>=0.15,<0.17
